Added separate change batches for alias target records
When committing changes that include records that are aliases to other target
records, those targets must already exist in the zone, and they can't be
part of the same ResourceRecordSets change batch, as Route53 will reject the
whole update.

Implemented a logic to identify the alias target records that don't yet exist
in the zone, and group them into a separate ResourceRecordSets change batch
that is executed before the main zone update.